Integrate with fixer.io

+++ solution/src/main/java/uk/co/engagetech/challenge/domain/Expense.java
@@ -24,7 +24,7 @@ public class Expense {
     private Date date;
 
     @JsonDeserialize(using = MoneyDeserializer.class)
-    private BigInteger amount;
+    private Long amount;
 
     private String reason;
